+++ src/styles/Comment/CommentsForm.module.scss
@@ -1,4 +1,4 @@
-@import "../App/App.module.scss";
+@import "../Common/Common.module.scss";
 
 .form {
   @extend .card;
