+++ src/styles/Common/ColorPicker.module.scss
