+++ src/styles/Follow/Follow.module.scss
@@ -1,4 +1,4 @@
-@import "../App/App.module.scss";
+@import "../Common/Common.module.scss";
 
 .follow {
   display: flex;
