+++ src/styles/Group/ItemAvatars.module.scss
@@ -1,4 +1,4 @@
-@import "../App/colors.scss";
+@import "../Common/colors.scss";
 
 .avatars {
   display: flex;
