+++ src/styles/Group/ToggleForms.module.scss
@@ -1,4 +1,4 @@
-@import "../App/App.module.scss";
+@import "../Common/Common.module.scss";
 
 .toggleContainer {
   @extend .card;
