+++ src/styles/Motion/ActionFields.module.scss
@@ -1,4 +1,4 @@
-@import "../App/App.module.scss";
+@import "../Common/Common.module.scss";
 
 .imageInputIcon {
   margin-top: 5px;
