+++ src/styles/Motion/Motion.module.scss
@@ -1,5 +1,5 @@
-@import "../App/App.module.scss";
-@import "../App/colors.scss";
+@import "../Common/Common.module.scss";
+@import "../Common/colors.scss";
 
 .info {
   color: $lightGray;
