+++ src/styles/Welcome/Welcome.module.scss
@@ -1,1 +1,1 @@
-@import "../App/App.module.scss";
+@import "../Common/Common.module.scss";
